Se agrega la conexion a la base de datos

+++ EPATEC/Models/Usuario.cs
@@ -12,7 +12,7 @@ namespace EPATEC.Models
         public long Id_usuario { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int Penalizacion { get; set; }
+        public short Penalizacion { get; set; }
         public string Residencia { get; set; }
         public string Fecha_de_Nacimiento { get; set; }
         public int Telefono { get; set; }
